requirements.txt
fastapi
uvicorn[standard]
pydantic
pyjwt
pytest
pytest-cov
httpx
flake8
python-dotenv
